+++ ext/d_heap/d_heap.c
@@ -40,6 +40,9 @@ struct dheap_struct
     long   size;
     long   capa;
     ENTRY *entries;
+#ifdef DHEAP_MAP
+    VALUE indexes; // Hash
+#endif
 };
 
 struct dheap_entry
@@ -47,6 +50,8 @@ struct dheap_entry
     SCORE score;
     VALUE value;
 };
+
+#define DHEAPMAP_P(heap) UNLIKELY(RTEST((heap)->indexes))
 
 /********************************************************************
  *
@@ -63,13 +68,44 @@ struct dheap_entry
 #define DHEAP_MAX_CAPA      (LONG_MAX / (int)sizeof(ENTRY))
 #define DHEAP_CAPA_INCR_MAX (10 * 1024 * 1024 / (int)sizeof(ENTRY))
 
-static ID id_cmp;         // <=>
-static ID id_abs;         // abs
-static ID id_lshift;      // <<
-static ID id_unary_minus; // -@
+static ID id_cmp;    // <=>
+static ID id_abs;    // abs
+static ID id_lshift; // <<
+static ID id_uminus; // -@
 
-static const ENTRY EmptyDheapEntry; // 0 value for safety overwrite after pop
 static const rb_data_type_t dheap_data_type;
+
+/********************************************************************
+ *
+ * Metaprogramming macros
+ *
+ ********************************************************************/
+
+#define LIKELY   RB_LIKELY
+#define UNLIKELY RB_UNLIKELY
+
+#ifdef DHEAP_MAP
+#    define DHEAP_DISPATCH_EXPR(func, heap, ...)                               \
+        (DHEAPMAP_P(heap) ? dheapmap_##func(heap, __VA_ARGS__)                 \
+                          : dheap_##func(heap, __VA_ARGS__))
+#else
+#    define DHEAP_DISPATCH_EXPR(func, heap, ...)                               \
+        dheap_##func(heap, __VA_ARGS__);
+#endif
+
+#ifdef DHEAP_MAP
+#    define DHEAP_DISPATCH_STMT(heap, macro, ...)                              \
+        do {                                                                   \
+            if (DHEAPMAP_P(heap)) {                                            \
+                macro(dheapmap, heap, __VA_ARGS__);                            \
+            } else {                                                           \
+                macro(dheap, heap, __VA_ARGS__);                               \
+            }                                                                  \
+        } while (0)
+#else
+#    define DHEAP_DISPATCH_STMT(heap, macro, ...)                              \
+        macro(dheap, heap, __VA_ARGS__)
+#endif
 
 /********************************************************************
  *
@@ -86,7 +122,7 @@ SCORE2NUM(SCORE s)
     if (roundl(s) == s) {
         if (s < 0) {
             unsigned long long ull = (unsigned long long)(-s);
-            return rb_funcall(ULL2NUM(ull), id_unary_minus, 0, Qundef);
+            return rb_funcall(ULL2NUM(ull), id_uminus, 0, Qundef);
         }
         return ULL2NUM((unsigned long long)(s));
     }
@@ -170,18 +206,29 @@ VAL2SCORE(VALUE score)
  *
  ********************************************************************/
 
-#define DHEAP_SCORE(heap, idx) ((heap)->entries[idx].score)
-#define DHEAP_VALUE(heap, idx) ((heap)->entries[idx].value)
+#define IDX_IN_RANGE(heap, idx) ((idx) < 0 || (heap)->size <= (idx))
+#define DHEAP_SCORE(heap, idx)  (DHEAP_GET(heap, idx).score)
+#define DHEAP_VALUE(heap, idx)  (DHEAP_GET(heap, idx).value)
 
-static inline VALUE
-DHEAP_ENTRY_ARY(dheap_t *heap, long idx)
-{
-    if (idx < 0 || heap->size <= idx) {
-        return Qnil;
-    }
-    return rb_ary_new_from_args(
-      2, DHEAP_VALUE(heap, 0), SCORE2NUM(DHEAP_SCORE(heap, 0)));
-}
+#define DHEAP_ENTRY_ARY(heap, idx)                                             \
+    (IDX_IN_RANGE(heap, idx)                                                   \
+       ? Qnil                                                                  \
+       : rb_ary_new_from_args(                                                 \
+           2, DHEAP_VALUE(heap, idx), SCORE2NUM(DHEAP_SCORE(heap, idx))))
+
+#define DHEAP_GET(heap, idx) ((heap)->entries[idx])
+#define DHEAP_SET(T, heap, index, entry)                                       \
+    do {                                                                       \
+        DHEAP_GET(heap, index) = (entry);                                      \
+        DHEAP_SET_##T(heap, index, entry);                                     \
+    } while (0)
+
+#define DHEAP_SET_dheap(heap, index, entry) /* noop */
+
+#ifdef DHEAP_MAP
+#    define DHEAP_SET_dheapmap(heap, index, entry)                             \
+        rb_hash_aset((heap)->indexes, (entry).value, LONG2NUM(index))
+#endif
 
 /********************************************************************
  *
@@ -219,8 +266,12 @@ dheap_compact(void *ptr)
 {
     dheap_t *heap = ptr;
     for (long i = 0; i < heap->size; ++i) {
-        if (DHEAP_VALUE(heap, i)) rb_gc_location(DHEAP_VALUE(heap, i));
+        if (DHEAP_VALUE(heap, i))
+            DHEAP_VALUE(heap, i) = rb_gc_location(DHEAP_VALUE(heap, i));
     }
+#    ifdef DHEAP_MAP
+    if (DHEAPMAP_P(heap)) heap->indexes = rb_gc_location(heap->indexes);
+#    endif
 }
 #else
 #    define rb_gc_mark_movable(x) rb_gc_mark(x)
@@ -233,6 +284,9 @@ dheap_mark(void *ptr)
     for (long i = 0; i < heap->size; ++i) {
         if (DHEAP_VALUE(heap, i)) rb_gc_mark_movable(DHEAP_VALUE(heap, i));
     }
+#ifdef DHEAP_MAP
+    if (DHEAPMAP_P(heap)) rb_gc_mark_movable(heap->indexes);
+#endif
 }
 
 static void
@@ -246,6 +300,9 @@ dheap_free(void *ptr)
     }
     heap->capa = 0;
     xfree(ptr);
+#ifdef DHEAP_MAP
+    heap->indexes = Qnil;
+#endif
 }
 
 static size_t
@@ -345,6 +402,9 @@ dheap_s_alloc(VALUE klass)
     heap->size    = 0;
     heap->capa    = 0;
     heap->entries = NULL;
+#ifdef DHEAP_MAP
+    heap->indexes = Qnil;
+#endif
 
     return obj;
 }
@@ -355,6 +415,13 @@ get_dheap_struct(VALUE self)
     dheap_t *heap;
     TypedData_Get_Struct(self, dheap_t, &dheap_data_type, heap);
     return heap;
+}
+
+static inline dheap_t *
+get_dheap_struct_unfrozen(VALUE self)
+{
+    rb_check_frozen(self);
+    return get_dheap_struct(self);
 }
 
 void
@@ -401,12 +468,8 @@ static inline int
 dheap_value_to_int_d(VALUE num)
 {
     int d = NUM2INT(num);
-    if (d < 2) {
-        rb_raise(rb_eArgError, "DHeap d=%u is too small", d);
-    }
-    if (d > DHEAP_MAX_D) {
-        rb_raise(rb_eArgError, "DHeap d=%u is too large", d);
-    }
+    if (d < 2) rb_raise(rb_eArgError, "DHeap d=%u is too small", d);
+    if (d > DHEAP_MAX_D) rb_raise(rb_eArgError, "DHeap d=%u is too large", d);
     return d;
 }
 
@@ -421,7 +484,7 @@ dheap_value_to_long_capa(VALUE num)
 }
 
 static VALUE
-dheap_initialize(VALUE self, VALUE d, VALUE capa)
+dheap_init(VALUE self, VALUE d, VALUE capa, VALUE map)
 {
     dheap_t *heap = get_dheap_struct(self);
 
@@ -430,6 +493,9 @@ dheap_initialize(VALUE self, VALUE d, VALUE capa)
 
     heap->d = dheap_value_to_int_d(d);
     dheap_set_capa(heap, dheap_value_to_long_capa(capa));
+#ifdef DHEAP_MAP
+    if (RTEST(map)) heap->indexes = rb_hash_new();
+#endif
 
     return self;
 }
@@ -438,11 +504,8 @@ dheap_initialize(VALUE self, VALUE d, VALUE capa)
 static VALUE
 dheap_initialize_copy(VALUE copy, VALUE orig)
 {
-    dheap_t *heap_copy;
+    dheap_t *heap_copy = get_dheap_struct_unfrozen(copy);
     dheap_t *heap_orig = get_dheap_struct(orig);
-
-    rb_check_frozen(copy);
-    TypedData_Get_Struct(copy, dheap_t, &dheap_data_type, heap_copy);
 
     heap_copy->d = heap_orig->d;
 
@@ -450,6 +513,10 @@ dheap_initialize_copy(VALUE copy, VALUE orig)
     heap_copy->size = heap_orig->size;
     if (heap_copy->size)
         MEMCPY(heap_copy->entries, heap_orig->entries, ENTRY, heap_orig->size);
+#ifdef DHEAP_MAP
+    if (RTEST(heap_orig->indexes))
+        heap_copy->indexes = rb_hash_dup(heap_orig->indexes);
+#endif
 
     return copy;
 }
@@ -460,40 +527,25 @@ dheap_initialize_copy(VALUE copy, VALUE orig)
  *
  ********************************************************************/
 
-VALUE
-dheap_sift_up(dheap_t *heap, long index)
-{
-    ENTRY entry = heap->entries[index];
+#define DHEAP_SIFT_UP(T, heap, i)                                              \
+    do {                                                                       \
+        long  sift_idx = i;                                                    \
+        ENTRY entry    = DHEAP_GET(heap, sift_idx);                            \
+        ASSERT_DHEAP_IDX_OK(heap, sift_idx);                                   \
+        for (long parent_idx; 0 < sift_idx; sift_idx = parent_idx) {           \
+            parent_idx = DHEAP_IDX_PARENT(heap, sift_idx);                     \
+            if (CMP_LTE(DHEAP_SCORE((heap), parent_idx), entry.score)) break;  \
+            DHEAP_SET(T, heap, sift_idx, DHEAP_GET(heap, parent_idx));         \
+        }                                                                      \
+        DHEAP_SET(T, heap, sift_idx, entry);                                   \
+    } while (0)
 
-    ASSERT_DHEAP_IDX_OK(heap, index);
-
-    // sift it up to where it belongs
-    for (long parent_index; 0 < index; index = parent_index) {
-        /* debug(rb_sprintf("sift up(%"PRIsVALUE", %d, %ld)", heap->values,
-         * heap->d, index)); */
-        parent_index = DHEAP_IDX_PARENT(heap, index);
-
-        // parent is smaller: heap is restored
-        if (CMP_LTE(DHEAP_SCORE(heap, parent_index), entry.score)) break;
-
-        // parent is larger: swap and continue sifting up
-        heap->entries[index] = heap->entries[parent_index];
-    }
-    heap->entries[index] = entry;
-    /* debug(rb_sprintf("sifted (%"PRIsVALUE", %d, %ld)", heap->values, heap->d,
-     * index)); */
-    return LONG2NUM(index);
-}
-
-/*
- * this is a tiny bit more complicated than the binary heap version
- */
 static inline long
-dheap_find_min_child(dheap_t *heap, long parent, long last_index)
+dheap_min_child(dheap_t *heap, long parent, long last_index)
 {
     long min_child = DHEAP_IDX_CHILD_0(heap, parent);
     long last_sib  = DHEAP_IDX_CHILD_D(heap, parent);
-    if (last_index < last_sib) last_sib = last_index;
+    if (UNLIKELY(last_index < last_sib)) last_sib = last_index;
 
     for (long sibidx = min_child + 1; sibidx <= last_sib; ++sibidx) {
         if (CMP_LT(DHEAP_SCORE(heap, sibidx), DHEAP_SCORE(heap, min_child))) {
@@ -503,36 +555,26 @@ dheap_find_min_child(dheap_t *heap, long parent, long last_index)
     return min_child;
 }
 
-void
-dheap_sift_down(dheap_t *heap, long index, long last_index)
-{
-    if (last_index < 1 || DHEAP_IDX_PARENT(heap, last_index) < index) {
-        // short-circuit: no chance for a child
-        return;
-    } else {
-        ENTRY entry = heap->entries[index];
+#define DHEAP_CAN_SIFT_DOWN(heap, index, last_index)                           \
+    (LIKELY(1 <= last_index || index <= DHEAP_IDX_PARENT(heap, last_index)))
 
-        long last_parent = DHEAP_IDX_PARENT(heap, last_index);
-
-        ASSERT_DHEAP_IDX_OK(heap, index);
-
-        // iteratively sift it down to where it belongs
-        while (index <= last_parent) {
-            // find min child
-            long min_child = dheap_find_min_child(heap, index, last_index);
-
-            // child is larger: heap is restored
-            if (CMP_LTE(entry.score, DHEAP_SCORE(heap, min_child))) break;
-
-            // child is smaller: swap and continue sifting down
-            heap->entries[index] = heap->entries[min_child];
-            index                = min_child;
-        }
-        heap->entries[index] = entry;
-        // debug(rb_sprintf("sifted (%"PRIsVALUE", %d, %ld)", heap->values,
-        // heap->d, index));
-    }
-}
+#define DHEAP_SIFT_DOWN(T, heap, i)                                            \
+    do {                                                                       \
+        long sift_idx = i;                                                     \
+        long last_idx = DHEAP_IDX_LAST(heap);                                  \
+        ASSERT_DHEAP_IDX_OK(heap, sift_idx);                                   \
+        if (DHEAP_CAN_SIFT_DOWN(heap, sift_idx, last_idx)) {                   \
+            ENTRY entry       = heap->entries[sift_idx];                       \
+            long  last_parent = DHEAP_IDX_PARENT(heap, last_idx);              \
+            while (sift_idx <= last_parent) {                                  \
+                long min_child = dheap_min_child(heap, sift_idx, last_idx);    \
+                if (CMP_LTE(entry.score, DHEAP_SCORE(heap, min_child))) break; \
+                DHEAP_SET(T, heap, sift_idx, (heap)->entries[min_child]);      \
+                sift_idx = min_child;                                          \
+            }                                                                  \
+            DHEAP_SET(T, heap, sift_idx, entry);                               \
+        }                                                                      \
+    } while (0)
 
 /********************************************************************
  *
@@ -550,6 +592,8 @@ dheap_size(VALUE self)
     return LONG2NUM(heap->size);
 }
 
+#define DHEAP_EMPTY_P(heap) UNLIKELY((heap)->size <= 0)
+
 /*
  * @return [Boolean] if the heap is empty
  */
@@ -557,7 +601,7 @@ static VALUE
 dheap_empty_p(VALUE self)
 {
     dheap_t *heap = get_dheap_struct(self);
-    return heap->size ? Qfalse : Qtrue;
+    return DHEAP_EMPTY_P(heap) ? Qtrue : Qfalse;
 }
 
 /*
@@ -576,14 +620,49 @@ dheap_attr_d(VALUE self)
  *
  ********************************************************************/
 
+#define DHEAP_PUSH(T, heap, entry)                                             \
+    do {                                                                       \
+        dheap_ensure_room_for_push(heap, 1);                                   \
+        DHEAP_SET(T, heap, (heap)->size, *(entry));                            \
+        ++heap->size;                                                          \
+        DHEAP_SIFT_UP(T, heap, DHEAP_IDX_LAST(heap));                          \
+    } while (0)
+
 static inline void
-dheap_push_entry(dheap_t *heap, ENTRY *entry)
+dheap_push_entry(VALUE self, ENTRY *entry)
 {
-    dheap_ensure_room_for_push(heap, 1);
-    heap->entries[heap->size] = *entry;
-    ++heap->size;
-    dheap_sift_up(heap, heap->size - 1);
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    DHEAP_PUSH(dheap, heap, entry);
 }
+
+#ifdef DHEAP_MAP
+static inline long
+dheapmap_index_of(dheap_t *heap, VALUE value);
+
+static inline void
+dheapmap_update_entry(dheap_t *heap, long index, ENTRY *entry)
+{
+    SCORE prev = DHEAP_SCORE(heap, index);
+    DHEAP_SET(dheapmap, heap, index, *entry);
+    if (CMP_LT(prev, entry->score)) {
+        DHEAP_SIFT_DOWN(dheapmap, heap, index);
+    } else {
+        DHEAP_SIFT_UP(dheapmap, heap, index);
+    }
+}
+
+static inline void
+dheapmap_push_entry(VALUE self, ENTRY *entry)
+{
+    dheap_t *heap  = get_dheap_struct_unfrozen(self);
+    long     index = dheapmap_index_of(heap, entry->value);
+    if (0 <= index) {
+        dheapmap_update_entry(heap, index, entry);
+        return;
+    }
+    DHEAP_PUSH(dheapmap, heap, entry);
+}
+#endif
 
 /*
  * Inserts a value into the heap, using a score to determine sort-order.
@@ -600,15 +679,30 @@ dheap_push_entry(dheap_t *heap, ENTRY *entry)
 static VALUE
 dheap_insert(VALUE self, VALUE score, VALUE value)
 {
-    ENTRY    entry;
-    dheap_t *heap = get_dheap_struct(self);
-    rb_check_frozen(self);
-
-    entry.score = VAL2SCORE(score);
-    entry.value = value;
-    dheap_push_entry(heap, &entry);
-
+    ENTRY entry = { VAL2SCORE(score), value };
+    dheap_push_entry(self, &entry);
     return self;
+}
+
+#ifdef DHEAP_MAP
+/* (see DHeap#insert) */
+static VALUE
+dheapmap_insert(VALUE self, VALUE score, VALUE value)
+{
+    ENTRY entry = { VAL2SCORE(score), value };
+    dheapmap_push_entry(self, &entry);
+    return self;
+}
+#endif
+
+static inline ENTRY
+dheap_push_args_to_entry(int argc, VALUE *argv)
+{
+    ENTRY entry;
+    rb_check_arity(argc, 1, 2);
+    entry.value = argv[0];
+    entry.score = VAL2SCORE(argc < 2 ? entry.value : argv[1]);
+    return entry;
 }
 
 /*
@@ -631,17 +725,21 @@ dheap_insert(VALUE self, VALUE score, VALUE value)
 static VALUE
 dheap_push(int argc, VALUE *argv, VALUE self)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    ENTRY    entry;
-    rb_check_frozen(self);
-
-    rb_check_arity(argc, 1, 2);
-    entry.value = argv[0];
-    entry.score = VAL2SCORE(argc < 2 ? entry.value : argv[1]);
-    dheap_push_entry(heap, &entry);
-
+    ENTRY entry = dheap_push_args_to_entry(argc, argv);
+    dheap_push_entry(self, &entry);
     return self;
 }
+
+#ifdef DHEAP_MAP
+/* (see DHeap#push) */
+static VALUE
+dheapmap_push(int argc, VALUE *argv, VALUE self)
+{
+    ENTRY entry = dheap_push_args_to_entry(argc, argv);
+    dheapmap_push_entry(self, &entry);
+    return self;
+}
+#endif
 
 /*
  * Pushes a value onto the heap.
@@ -655,18 +753,23 @@ dheap_push(int argc, VALUE *argv, VALUE self)
  * @return [self]
  */
 static VALUE
-dheap_left_shift(VALUE self, VALUE value)
+dheap_lshift(VALUE self, VALUE value)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    ENTRY    entry;
-    rb_check_frozen(self);
-
-    entry.score = VAL2SCORE(value);
-    entry.value = value;
-    dheap_push_entry(heap, &entry);
-
+    ENTRY entry = { VAL2SCORE(value), value };
+    dheap_push_entry(self, &entry);
     return self;
 }
+
+#ifdef DHEAP_MAP
+/* (see DHeap#<<) */
+static VALUE
+dheapmap_lshift(VALUE self, VALUE value)
+{
+    ENTRY entry = { VAL2SCORE(value), value };
+    dheapmap_push_entry(self, &entry);
+    return self;
+}
+#endif
 
 /********************************************************************
  *
@@ -674,23 +777,15 @@ dheap_left_shift(VALUE self, VALUE value)
  *
  ********************************************************************/
 
-static inline void
-dheap_del0(dheap_t *heap)
-{
-    if (0 < --heap->size) {
-        heap->entries[0]          = heap->entries[heap->size];
-        heap->entries[heap->size] = EmptyDheapEntry; // unnecessary to zero?
-        dheap_sift_down(heap, 0, heap->size - 1);
-    }
-}
-
-static inline VALUE
-dheap_pop0(dheap_t *heap)
-{
-    VALUE popped = DHEAP_VALUE(heap, 0);
-    dheap_del0(heap);
-    return popped;
-}
+#define PEEK_VALUE(heap)            DHEAP_VALUE(heap, 0)
+#define PEEK_SCORE(heap)            DHEAP_SCORE(heap, 0)
+#define PEEK_WITH_SCORE(heap)       DHEAP_ENTRY_ARY(heap, 0)
+#define PEEK_LT_P(heap, max_score)  _PEEK_CMP_P(heap, CMP_LT, max_score)
+#define PEEK_LTE_P(heap, max_score) _PEEK_CMP_P(heap, CMP_LTE, max_score)
+#define _PEEK_CMP_P(heap, cmp, score)                                          \
+    ((!DHEAP_EMPTY_P(heap) && cmp(PEEK_SCORE(heap), (score)))                  \
+       ? PEEK_VALUE(heap)                                                      \
+       : 0)
 
 /*
  * Returns the next value on the heap, and its score, without popping it
@@ -705,7 +800,7 @@ static VALUE
 dheap_peek_with_score(VALUE self)
 {
     dheap_t *heap = get_dheap_struct(self);
-    return DHEAP_ENTRY_ARY(heap, 0);
+    return PEEK_WITH_SCORE(heap);
 }
 
 /*
@@ -722,8 +817,8 @@ static VALUE
 dheap_peek_score(VALUE self)
 {
     dheap_t *heap = get_dheap_struct(self);
-    if (DHEAP_IDX_LAST(heap) < 0) return Qnil;
-    return SCORE2NUM(DHEAP_SCORE(heap, 0));
+    if (DHEAP_EMPTY_P(heap)) return Qnil;
+    return SCORE2NUM(PEEK_SCORE(heap));
 }
 
 /*
@@ -739,8 +834,51 @@ static VALUE
 dheap_peek(VALUE self)
 {
     dheap_t *heap = get_dheap_struct(self);
-    return heap->size ? DHEAP_VALUE(heap, 0) : Qnil;
+    if (DHEAP_EMPTY_P(heap)) return Qnil;
+    return PEEK_VALUE(heap);
 }
+
+#define DHEAP_POP_COMMON_INIT(self)                                            \
+    dheap_t *heap = get_dheap_struct_unfrozen(self);                           \
+    if (DHEAP_EMPTY_P(heap)) return Qnil;
+
+#define DHEAP_DELETE_0(T, heap)                                                \
+    do {                                                                       \
+        _DELETE_ENTRY(T, heap, 0);                                             \
+        if (0 < --(heap)->size) {                                              \
+            DHEAP_SET(T, (heap), 0, (heap)->entries[(heap)->size]);            \
+            DHEAP_SIFT_DOWN(T, (heap), 0);                                     \
+        }                                                                      \
+    } while (0)
+
+#define _DELETE_ENTRY(T, heap, idx)    _DELETE_ENTRY_##T(heap, idx)
+#define _DELETE_ENTRY_dheap(heap, idx) /* noop */
+#define _DELETE_ENTRY_dheapmap(heap, idx)                                      \
+    rb_hash_delete(heap->indexes, DHEAP_VALUE(heap, idx))
+
+#define POP(T, heap, popped)            _POP(T, VALUE, heap, popped)
+#define POP_WITH_SCORE(T, heap, popped) _POP(T, WITH_SCORE, heap, popped)
+#define POP_LT(T, heap, max, popped)    _POP_CMP(T, heap, LT, max, popped)
+#define POP_LTE(T, heap, max, popped)   _POP_CMP(T, heap, LTE, max, popped)
+
+#define _POP_CMP(T, heap, cmp, cmp_score, popped)                              \
+    do {                                                                       \
+        if ((*(popped) = PEEK_##cmp##_P(heap, cmp_score))) {                   \
+            DHEAP_DELETE_0(T, heap);                                           \
+        } else {                                                               \
+            *(popped) = Qnil;                                                  \
+        }                                                                      \
+    } while (0)
+
+#define _POP(T, peek_type, heap, popped)                                       \
+    do {                                                                       \
+        if (DHEAP_EMPTY_P(heap)) {                                             \
+            *(popped) = Qnil;                                                  \
+        } else {                                                               \
+            *(popped) = PEEK_##peek_type(heap);                                \
+            DHEAP_DELETE_0(T, heap);                                           \
+        }                                                                      \
+    } while (0)
 
 /*
  * Pops the minimum value from the top of the heap
@@ -757,10 +895,23 @@ dheap_peek(VALUE self)
 static VALUE
 dheap_pop(VALUE self)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    rb_check_frozen(self);
-    return heap->size ? dheap_pop0(heap) : Qnil;
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP(dheap, heap, &popped);
+    return popped;
 }
+
+#ifdef DHEAP_MAP
+/* (see DHeap#pop) */
+static VALUE
+dheapmap_pop(VALUE self)
+{
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP(dheapmap, heap, &popped);
+    return popped;
+}
+#endif
 
 /*
  * Pops the minimum value from the top of the heap, along with its score.
@@ -775,14 +926,26 @@ dheap_pop(VALUE self)
 static VALUE
 dheap_pop_with_score(VALUE self)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    VALUE    ary  = DHEAP_ENTRY_ARY(heap, 0);
-    rb_check_frozen(self);
-    if (ary != Qnil) {
-        dheap_pop0(heap);
-    }
-    return ary;
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP_WITH_SCORE(dheap, heap, &popped);
+    return popped;
 }
+
+#ifdef DHEAP_MAP
+/* (see DHeap#pop_with_score) */
+static VALUE
+dheapmap_pop_with_score(VALUE self)
+{
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP_WITH_SCORE(dheapmap, heap, &popped);
+    return popped;
+}
+#endif
+
+#define DHEAP_POP_IF(heap, cmp, max_score)                                     \
+    if (!cmp(PEEK_SCORE(heap), VAL2SCORE(max_score))) return Qnil
 
 /*
  * Pops the minimum value only if it is less than or equal to a max score.
@@ -801,12 +964,23 @@ dheap_pop_with_score(VALUE self)
 static VALUE
 dheap_pop_lte(VALUE self, VALUE max_score)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    rb_check_frozen(self);
-    if (heap->size <= 0) return Qnil;
-    if (!CMP_LTE(DHEAP_SCORE(heap, 0), VAL2SCORE(max_score))) return Qnil;
-    return dheap_pop0(heap);
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP_LTE(dheap, heap, VAL2SCORE(max_score), &popped);
+    return popped;
 }
+
+#ifdef DHEAP_MAP
+/* (see DHeap#pop_lte) */
+static VALUE
+dheapmap_pop_lte(VALUE self, VALUE max_score)
+{
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP_LTE(dheapmap, heap, VAL2SCORE(max_score), &popped);
+    return popped;
+}
+#endif
 
 /*
  * Pops the minimum value only if it is less than a max score.
@@ -825,33 +999,39 @@ dheap_pop_lte(VALUE self, VALUE max_score)
 static VALUE
 dheap_pop_lt(VALUE self, VALUE max_score)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    rb_check_frozen(self);
-    if (heap->size <= 0) return Qnil;
-    if (!CMP_LT(DHEAP_SCORE(heap, 0), VAL2SCORE(max_score))) return Qnil;
-    return dheap_pop0(heap);
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP_LT(dheap, heap, VAL2SCORE(max_score), &popped);
+    return popped;
 }
 
-#define DHEAP_PEEK_LT_P(heap, max_score)                                       \
-    (heap->size && CMP_LT(DHEAP_SCORE(heap, 0), max_score))
-
+#ifdef DHEAP_MAP
+/* (see DHeap#pop_lt) */
 static VALUE
-dheap_pop_all_below0(dheap_t *heap, SCORE max_score, VALUE array)
+dheapmap_pop_lt(VALUE self, VALUE max_score)
 {
-    if (!RTEST(array)) {
-        array = rb_ary_new();
-    }
-    if (RB_TYPE_P(array, T_ARRAY)) {
-        while (DHEAP_PEEK_LT_P(heap, max_score)) {
-            rb_ary_push(array, dheap_pop0(heap));
-        }
-    } else {
-        while (DHEAP_PEEK_LT_P(heap, max_score)) {
-            rb_funcall(array, id_lshift, 1, dheap_pop0(heap));
-        }
-    }
-    return array;
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    VALUE    popped;
+    POP_LT(dheapmap, heap, VAL2SCORE(max_score), &popped);
+    return popped;
 }
+#endif
+
+#define POP_ALL_BELOW(T, heap, max_score, array)                               \
+    do {                                                                       \
+        VALUE val = Qnil;                                                      \
+        if (RB_TYPE_P(array, T_ARRAY)) {                                       \
+            while ((val = PEEK_LT_P(heap, max_score))) {                       \
+                DHEAP_DELETE_0(T, heap);                                       \
+                rb_ary_push(array, val);                                       \
+            }                                                                  \
+        } else {                                                               \
+            while ((val = PEEK_LT_P(heap, max_score))) {                       \
+                DHEAP_DELETE_0(T, heap);                                       \
+                rb_funcall(array, id_lshift, 1, val);                          \
+            }                                                                  \
+        }                                                                      \
+    } while (0)
 
 /*
  * @overload pop_all_below(max_score, receiver = [])
@@ -872,18 +1052,12 @@ dheap_pop_all_below0(dheap_t *heap, SCORE max_score, VALUE array)
 static VALUE
 dheap_pop_all_below(int argc, VALUE *argv, VALUE self)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    SCORE    max_score;
-    VALUE    array;
-    rb_check_frozen(self);
+    dheap_t *heap      = get_dheap_struct_unfrozen(self);
+    SCORE    max_score = (argc) ? VAL2SCORE(argv[0]) : 0.0;
+    VALUE    array     = (argc == 1) ? rb_ary_new() : argv[1];
     rb_check_arity(argc, 1, 2);
-    max_score = VAL2SCORE(argv[0]);
-    if (argc == 1) {
-        array = rb_ary_new();
-    } else {
-        array = argv[1];
-    }
-    return dheap_pop_all_below0(heap, max_score, array);
+    DHEAP_DISPATCH_STMT(heap, POP_ALL_BELOW, max_score, array);
+    return array;
 }
 
 /********************************************************************
@@ -891,6 +1065,17 @@ dheap_pop_all_below(int argc, VALUE *argv, VALUE self)
  * DHeap, misc methods
  *
  ********************************************************************/
+
+static VALUE
+dheap_to_a(VALUE self)
+{
+    dheap_t *heap  = get_dheap_struct(self);
+    VALUE    array = rb_ary_new_capa(heap->size);
+    for (long i = 0; i < heap->size; i++) {
+        rb_ary_push(array, DHEAP_ENTRY_ARY(heap, i));
+    }
+    return array;
+}
 
 /*
  * Clears all values from the heap, leaving it empty.
@@ -900,13 +1085,67 @@ dheap_pop_all_below(int argc, VALUE *argv, VALUE self)
 static VALUE
 dheap_clear(VALUE self)
 {
-    dheap_t *heap = get_dheap_struct(self);
-    rb_check_frozen(self);
-    if (0 < heap->size) {
+    dheap_t *heap = get_dheap_struct_unfrozen(self);
+    if (!DHEAP_EMPTY_P(heap)) {
         heap->size = 0;
+#ifdef DHEAP_MAP
+        if (DHEAPMAP_P(heap)) rb_hash_clear(heap->indexes);
+#endif
     }
     return self;
 }
+
+/********************************************************************
+ *
+ * DHeap::Map methods
+ *
+ ********************************************************************/
+
+#ifdef DHEAP_MAP
+
+// returns -1 if it isn't found
+static inline long
+dheapmap_index_of(dheap_t *heap, VALUE value)
+{
+    VALUE index = rb_hash_aref(heap->indexes, value);
+    return RTEST(index) ? NUM2LONG(index) : -1;
+}
+
+/*
+ * Retrieves the score that has been assigned to a heap member.
+ *
+ * Time complexity: <b>O(1)</b>
+ *
+ * @param object [Object] an object to lookup
+ * @return [Float,Integer,nil] the score associated with the object,
+ *                             or nil if the object isn't a member
+ */
+static VALUE
+dheapmap_aref(VALUE self, VALUE object)
+{
+    dheap_t *heap  = get_dheap_struct(self);
+    long     index = dheapmap_index_of(heap, object);
+    return (0 <= index) ? SCORE2NUM(DHEAP_SCORE(heap, index)) : Qnil;
+}
+
+/*
+ * Assign a score to an object, adding it to the heap or updating as necessary.
+ *
+ * Time complexity: <b>O(log n)</b> (score decrease will be faster than
+ * score increase)
+ *
+ * @param object [Object] an object to lookup
+ * @param score [Integer,#to_f] the score to set
+ * @return [Float,Integer] the score
+ */
+static VALUE
+dheapmap_aset(VALUE self, VALUE object, VALUE score)
+{
+    dheapmap_insert(self, score, object);
+    return score;
+}
+
+#endif
 
 /********************************************************************
  *
@@ -918,13 +1157,25 @@ void
 Init_d_heap(void)
 {
     VALUE rb_cDHeap = rb_define_class("DHeap", rb_cObject);
+#ifdef DHEAP_MAP
+    VALUE rb_cDHeapMap = rb_define_class_under(rb_cDHeap, "Map", rb_cDHeap);
+#endif
 
-    id_cmp         = rb_intern_const("<=>");
-    id_abs         = rb_intern_const("abs");
-    id_lshift      = rb_intern_const("<<");
-    id_unary_minus = rb_intern_const("-@");
+    id_cmp    = rb_intern_const("<=>");
+    id_abs    = rb_intern_const("abs");
+    id_lshift = rb_intern_const("<<");
+    id_uminus = rb_intern_const("-@");
 
     rb_define_alloc_func(rb_cDHeap, dheap_s_alloc);
+
+#ifdef DHEAP_MAP
+#    define def_override_inherited(rb_name, c_name, argc)                      \
+        rb_define_method(rb_cDHeap, rb_name, dheap_##c_name, argc);            \
+        rb_define_method(rb_cDHeapMap, rb_name, dheapmap_##c_name, argc);
+#else
+#    define def_override_inherited(rb_name, c_name, argc)                      \
+        rb_define_method(rb_cDHeap, rb_name, dheap_##c_name, argc);
+#endif
 
     /*
      * This is based on INT_MAX. But it is very very unlikely you will want a
@@ -946,26 +1197,31 @@ Init_d_heap(void)
      */
     rb_define_const(rb_cDHeap, "DEFAULT_CAPA", INT2NUM(DHEAP_DEFAULT_CAPA));
 
-    rb_define_private_method(
-      rb_cDHeap, "__init_without_kw__", dheap_initialize, 2);
+    rb_define_private_method(rb_cDHeap, "__init_without_kw__", dheap_init, 3);
     rb_define_method(rb_cDHeap, "initialize_copy", dheap_initialize_copy, 1);
 
     rb_define_method(rb_cDHeap, "d", dheap_attr_d, 0);
     rb_define_method(rb_cDHeap, "size", dheap_size, 0);
     rb_define_method(rb_cDHeap, "empty?", dheap_empty_p, 0);
-    rb_define_method(rb_cDHeap, "peek", dheap_peek, 0);
+    rb_define_method(rb_cDHeap, "to_a", dheap_to_a, 0);
 
     rb_define_method(rb_cDHeap, "clear", dheap_clear, 0);
-    rb_define_method(rb_cDHeap, "insert", dheap_insert, 2);
-    rb_define_method(rb_cDHeap, "push", dheap_push, -1);
-    rb_define_method(rb_cDHeap, "<<", dheap_left_shift, 1);
-    rb_define_method(rb_cDHeap, "pop", dheap_pop, 0);
-
-    rb_define_method(rb_cDHeap, "pop_lt", dheap_pop_lt, 1);
-    rb_define_method(rb_cDHeap, "pop_lte", dheap_pop_lte, 1);
-    rb_define_method(rb_cDHeap, "pop_all_below", dheap_pop_all_below, -1);
-
+    rb_define_method(rb_cDHeap, "peek", dheap_peek, 0);
     rb_define_method(rb_cDHeap, "peek_score", dheap_peek_score, 0);
     rb_define_method(rb_cDHeap, "peek_with_score", dheap_peek_with_score, 0);
-    rb_define_method(rb_cDHeap, "pop_with_score", dheap_pop_with_score, 0);
+    rb_define_method(rb_cDHeap, "pop_all_below", dheap_pop_all_below, -1);
+
+    def_override_inherited("insert", insert, 2);
+    def_override_inherited("push", push, -1);
+    def_override_inherited("<<", lshift, 1);
+
+    def_override_inherited("pop", pop, 0);
+    def_override_inherited("pop_lt", pop_lt, 1);
+    def_override_inherited("pop_lte", pop_lte, 1);
+    def_override_inherited("pop_with_score", pop_with_score, 0);
+
+#ifdef DHEAP_MAP
+    rb_define_method(rb_cDHeapMap, "[]", dheapmap_aref, 1);
+    rb_define_method(rb_cDHeapMap, "[]=", dheapmap_aset, 2);
+#endif
 }